+++ blocks/core/ff_module/ff_module-task-response-actions/ff_module-task-response-actions.js
@@ -1,10 +1,10 @@
 'use strict';
 
 var React = require('react');
-var template = require('./_ff_module-task-event-actions.rt.js');
+var template = require('./_ff_module-task-response-actions.rt.js');
 
 module.exports = React.createClass({
-    displayName: 'TaskEventActions',
+    displayName: 'TaskResponseActions',
     render: template,
     propTypes: {
         onClick: React.PropTypes.func.isRequired
