+++ blocks/core/ff_module/ff_module-task-response-actions/ff_module-task-response-actions.less
@@ -1,4 +1,4 @@
-.ff_module-task-event-actions {
+.ff_module-task-response-actions {
     &__dropdown .ff_module-dropdown-button__list {
         min-width:15em;
     }
