+++ blocks/core/ff_module/ff_module-task-response-actions/ff_module-task-response-actions.md
@@ -1,4 +1,4 @@
-<div data-ff_module-task-event-actions=""></div>
+<div data-ff_module-task-response-actions=""></div>
 
 This component renders a collection of buttons that are used in the Task Responses screen. A single click handler is passed in, and each button returns a task event with a defined `type` property on that handler.
